Do not use f-strings in log calls.
It is in general bad practice; in particular if one want a structured
logging handler that store the template string and the values separately, using
f-strings makes it not possible.

I belive there is also a question of actually calling the formatter on
the interpolated values which is not done if the loging level is low
enough.

WHile it might not be too critical for micropip, I belive it is good
practive to show the example.