Has setup_logging not override already setup loging level.
Add a verbosity=None (and make it the default) for verbosity.
If one has already increase the logging level, the default value of
verbose should not change it. This is achieved with a default value of
None.

Technically I guess we should have a context manager to restaure
previous value if True/False is passed, but that will be for another
time.

I also think that if verbosity is an int, we should set the lgging level
to that int instead of havin an arbitrary verbosity of 0/1/2; but that's
also another discussion.

I'm doing this as I'd like adding some logger.debug, but I don't want to
pass verbose=True everywhere.

Restore logger level between calls

forward getattr/setattr

unconditional context manager