Differenctiate `ValueError`s
ValueErrors are a bit too generic, and by raising/catching them we loose
a potential useful information; For example an invalid content type for
the index, is dealt as the same way than a missing wheel.

This introduces new exceptions for various cases, ant stop catching the
now UnsupportedContentTypeError, as I believe it should be a hard error as
we are likely having either an error with the index software, or a user
configuraition error.

See also beginning of discussion in #142