Do not store pypi_response as gz.
This uncompress all files in `pypi_response` and update the relevant
section of test code to deal with the new format.

The idea is that binary blobs as test data are frowned upon for various
reasons, even more recenlty with the `xz` fiasco where there were used
to carry a backdoor.

Compression does not really reduce repository size as git is compressed
anyway;

And uncompressed files are harder to modify/review modifications.